requirements.txt
PyYAML
aiohttp
discord.py
openai
